Create queue infrastructure
This commit sets the stage for our queue system, which will live inside
of `src/server/queues`.  Some of the key infrastructure points:

* Queue factories: Queues in our system are created through
a QueueFactory, which must extend `AbstractQueueFactory` and override
the name and processor location.

* Queue schedulers: Whe want to regularly schedule certain queues when
in production.  All queues must have a schedule (the schedule can be
NONE). Schedules are defined in the Schedules constant in
`queues/index.js`

* Queue processors: As already established in a previous commit, our
queues will be processed using named processors.  The processor will
live in the same directory as the factory.

This also adds two new yarn scripts:

* `queue:scheduleJobs` will schedule the queues.
* `queue:unscheduleJobs` will unschedule the queues.
* `queue:listScheduledJobs` will list the currently scheduled jobs.

This also adds a new environment variable:

* REDIS_URL - this is the URL of the redis instance, which is necessary
for the queue to work.  If this is not set then bull currently defaults
to localhost and the default redis port.

Issue #20
Issue #43